python-openai/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.9.0
openai==1.99.9
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn==0.23.2
python-multipart==0.0.7
